--- a/peekingduck/nodes/dabble/trackingv1/tracking_files/iou_tracker.py
+++ b/peekingduck/nodes/dabble/trackingv1/tracking_files/iou_tracker.py
@@ -40,10 +40,8 @@ from typing import Any, Dict, List, Tuple
 import numpy as np
 
 from peekingduck.nodes.dabble.trackingv1.tracking_files.track import Track
-from peekingduck.nodes.dabble.trackingv1.tracking_files.utils import (
-    iou_candidates,
-    xyxyn2tlwh,
-)
+from peekingduck.nodes.dabble.trackingv1.tracking_files.utils import iou_candidates
+from peekingduck.utils.bbox.transforms import xyxyn2tlwh
 
 
 class IOUTracker: