--- a/peekingduck/nodes/dabble/zone_count.py
+++ b/peekingduck/nodes/dabble/zone_count.py
@@ -14,7 +14,9 @@
 
 """Counts the number of detected objects within a boundary."""
 
-from typing import Any, Dict, List, Union
+from typing import Any, Dict, List, Tuple, Union
+
+from shapely.geometry.polygon import Point
 
 from peekingduck.nodes.abstract_node import AbstractNode
 from peekingduck.nodes.dabble.zoningv1.zone import Zone
@@ -53,23 +55,22 @@ class Node(AbstractNode):
     def __init__(self, config: Dict[str, Any] = None, **kwargs: Any) -> None:
         super().__init__(config, node_path=__name__, **kwargs)
         self.zones = [self._create_zone(zone) for zone in self.zones]  # type: ignore
+        self.zone_polygon_points = [zone.polygon_points for zone in self.zones]
 
     def run(self, inputs: Dict[str, Any]) -> Dict[str, Any]:
         """Counts all detected objects that falls within any specified zone,
         and return the total object count in each zone.
         """
         zone_counts = [0] * len(self.zones)
+        points = [Point(point) for point in inputs["btm_midpoint"]]
 
         # for each x, y point, check if it is in any zone and add count
-        for point in inputs["btm_midpoint"]:
-            for i, zone in enumerate(self.zones):
+        for i, zone in enumerate(self.zones):
+            for point in points:
                 if zone.contains(point):
                     zone_counts[i] += 1
 
-        return {
-            "zones": [zone.polygon_points for zone in self.zones],
-            "zone_count": zone_counts,
-        }
+        return {"zones": self.zone_polygon_points, "zone_count": zone_counts}
 
     def _create_zone(self, zone: List[List[Union[float, int]]]) -> Zone:
         """Creates the appropriate Zone given either the absolute pixel values
@@ -77,14 +78,12 @@ class Node(AbstractNode):
         """
         if all(all(0 <= i <= 1 for i in coords) for coords in zone):
             # coordinates are in fraction. Use resolution to get correct coords
-            zone_points = [
-                self._get_pixel_coords(coords, self.resolution) for coords in zone
-            ]
+            zone_points = list(map(self._get_pixel_coords, zone))
         elif all(
             all((isinstance(i, int) and i >= 0) for i in coords) for coords in zone
         ):
             # list is in pixel value.
-            zone_points = zone  # type: ignore
+            zone_points = list(map(tuple, zone))  # type: ignore
         else:
             raise ValueError(
                 f"Zone {zone} needs to be all pixel-wise points or all fractions "
@@ -98,9 +97,6 @@ class Node(AbstractNode):
         """Returns dictionary mapping the node's config keys to respective types."""
         return {"resolution": List[int], "zones": List[List[List[Union[int, float]]]]}
 
-    @staticmethod
-    def _get_pixel_coords(
-        coords: List[Union[float, int]], resolution: List[int]
-    ) -> List[int]:
+    def _get_pixel_coords(self, coords: List[Union[float, int]]) -> Tuple[int, int]:
         """Returns the pixel position of the zone points."""
-        return [int(coords[0] * resolution[0]), int(coords[1] * resolution[1])]
+        return int(coords[0] * self.resolution[0]), int(coords[1] * self.resolution[1])