--- a/peekingduck/nodes/model/efficientdet_d04/efficientdet_files/constants.py
+++ b/peekingduck/nodes/model/efficientdet_d04/efficientdet_files/constants.py
@@ -1,0 +1,23 @@
+# Copyright 2022 AI Singapore
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Code of this file is mostly forked from
+# [@xuannianz](https://github.com/xuannianz))
+
+"""Constants used by EfficientDet."""
+
+import numpy as np
+
+IMG_MEAN = np.array((0.485, 0.456, 0.406))
+IMG_STD = np.array((0.229, 0.224, 0.225))